feat(stats): add Validators section to all-chains network metrics page
